+++ scripts/lichessTools-init.js
@@ -97,6 +97,8 @@ lichessTools.loadTool(LiChessTools.Tools.ActiveIcon);
 lichessTools.loadTool(LiChessTools.Tools.ImagePasting);
 lichessTools.loadTool(LiChessTools.Tools.RemoveChatLinkWarning);
 lichessTools.loadTool(LiChessTools.Tools.MchatOptions);
+lichessTools.loadTool(LiChessTools.Tools.FullScreen);
+lichessTools.loadTool(LiChessTools.Tools.OneClickMove);
 //lichessTools.loadTool(LiChessTools.Tools.DownloadGames);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);