+++ tools/boardImage/tool.js
@@ -54,10 +54,13 @@
       ctx.canvas.width=800;
       ctx.canvas.height=800;
       const board = $('div.main-board cg-board');
-      //let url=board.css('background-image')?.replace(/url\(["']?|["']?\)/g,'');
       const href=$(ev.target).attr('href');
-      const m=/theme=(\w+)(?:&|$)/.exec(href);
-      let url=lichess.asset.url('../images/board/'+m[1]+'.jpg');
+      const backgroundText=window.getComputedStyle(board[0], ':before').backgroundImage;
+      let url=/"(.*)"/.exec(backgroundText||'')?.[1];
+      if (!url) {
+        const theme=$('body').attr('data-board')||'maple';
+        url=lichess.asset.url('../images/board/'+theme+'.jpg');
+      }
       let img=await this.getImage(url);
       ctx.drawImage(img,0,0,800,800);
       const q=800/board.width();